🔨 One team 컴포넌트 서버사이드 렌더링 전환
