+++ src/app/page.module.scss
@@ -4,10 +4,6 @@
   margin-top: 58px;
 }
 
-.PgApplyButtonContainer {
-  margin-top: 32px;
-}
-
 .PgAboutButtonContainer {
   margin: 120px 0px 40px;
 }
