+++ src/app/page.tsx
@@ -1,8 +1,9 @@
-'use client';
-
 import styles from './page.module.scss';
 
+import { getParts } from '@/apis/part';
+
 import Typo from '@/components/Home/Typo';
+import LandingButton from '@/components/Home/LandingButton';
 import About from '@/components/Home/About';
 import OneTeam from '@/components/Home/OneTeam';
 import Schedule from '@/components/Home/Schedule';
@@ -10,14 +11,14 @@ import Project from '@/components/Home/Project';
 import AsanDonation from '@/components/common/donation/AsanDonation';
 import PageRouterButton from '@/components/Home/PageRouterButton';
 
-import Button from '@/components/common/button/Button';
 import CommonWrapper from '@/components/common/layout/CommonWrapper';
 
-export default function Home() {
-  const handleApplyButton = () => {
-    // toDo: 구글 폼 나온 후 수정 필요
-    window.open('https://prography.org/');
-  };
+async function getPartData() {
+  return getParts();
+}
+
+export default async function Home() {
+  const team = await getPartData();
 
   return (
     <>
@@ -25,11 +26,7 @@ export default function Home() {
         <div className={styles.PgTypoContainer}>
           <Typo />
         </div>
-        <div className={styles.PgApplyButtonContainer}>
-          <Button buttonSize="56" onClick={handleApplyButton}>
-            9기 지원
-          </Button>
-        </div>
+        <LandingButton />
         <div className={styles.PgAboutButtonContainer}>
           <PageRouterButton label="About" href="/about" />
         </div>
@@ -37,7 +34,7 @@ export default function Home() {
         <div className={styles.PgTeamButtonContainer}>
           <PageRouterButton label="One team" href="/about" />
         </div>
-        <OneTeam />
+        <OneTeam team={team} />
         <h3 className={styles.PgScheduleHead}>9th Schedule</h3>
         <Schedule />
         <div className={styles.PgProjectButtonContainer}>
