+++ src/components/Home/About.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { useEffect, useState } from 'react';
 import CountUp from 'react-countup';
