+++ src/components/Home/LandingButton.module.scss
@@ -1,0 +1,20 @@
+@import '@/components/common/button/Button.scss';
+
+.PgApplyButtonContainer {
+  margin-top: 32px;
+}
+
+@media (min-width: 1280px) {
+  .PgApplyButtonContainer {
+    margin-top: 40px;
+
+    button {
+      @include button-style(
+        24px !important,
+        32px !important,
+        24px !important,
+        48px !important
+      );
+    }
+  }
+}
