+++ src/components/Home/OneTeam.tsx
@@ -1,18 +1,6 @@
-import { useState, useEffect } from 'react';
 import styles from './OneTeam.module.scss';
-import { PartType, getParts } from '@/apis/part';
 
-const OneTeam = () => {
-  // toDo: api 교체 필요
-  const [team, setTeam] = useState<PartType[]>([]);
-
-  useEffect(() => {
-    const fetchData = async () => {
-      const data = await getParts();
-      setTeam(data);
-    };
-    fetchData();
-  }, []);
+const OneTeam = ({ team }) => {
   return (
     <>
       <p className={styles.PgTeamText}>
