+++ src/components/Home/Project.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import 'swiper/css';
 import './Project.scss';
